# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HillSpace API",
    "description": "Estate management backend — auth, listings, search, verification, escrow",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "required": ["firstName", "lastName", "email", "phone", "password"],
        "properties": {
          "firstName": { "type": "string", "example": "Ada" },
          "lastName": { "type": "string", "example": "Lovelace" },
          "email": { "type": "string", "format": "email", "example": "ada@hillspace.test" },
          "phone": { "type": "string", "example": "+2348012345678" },
          "password": { "type": "string", "minLength": 8, "example": "password123" },
          "role": {
            "type": "string",
            "enum": ["buyer", "seller", "agent", "admin"],
            "example": "buyer"
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "ada@hillspace.test" },
          "password": { "type": "string", "example": "password123" }
        }
      },
      "RefreshTokenDto": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string" }
        }
      },
...
      "CreateEscrowDto": {
        "type": "object",
        "required": ["listingId", "amount"],
        "properties": {
          "listingId": { "type": "string", "example": "665f1a2b3c4d5e6f7a8b9c0d" },
          "amount": { "type": "number", "example": 85000000 },
          "currency": { "type": "string", "example": "NGN" }
        }
      },
      "FundEscrowDto": {
        "type": "object",
        "required": ["fundingReference"],
        "properties": {
          "fundingReference": { "type": "string", "example": "TRF-20260711-001" },
          "note": { "type": "string" }
        }
      },
      "TransitionEscrowDto": {
        "type": "object",
        "properties": {
          "note": { "type": "string" }
        }
      },
      "DisputeEscrowDto": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 10,
            "example": "Property condition does not match listing photos"
          }
        }
      },
...
      "AuthTokensResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "email": { "type": "string" },
              "role": { "type": "string" }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokensResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and receive JWT tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokensResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokensResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout and invalidate refresh token",
        "security": [{ "access-token": [] }],
        "responses": {
          "200": { "description": "Logged out" }
        }
      }
    },
...
    "/api/escrow": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Create an escrow deal on a listing",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEscrowDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Escrow created" }
        }
      }
    },
    "/api/escrow/me": {
      "get": {
        "tags": ["Escrow"],
        "summary": "List my escrow deals",
        "security": [{ "access-token": [] }],
        "responses": {
          "200": { "description": "My deals" }
        }
      }
    },
    "/api/escrow/{id}": {
      "get": {
        "tags": ["Escrow"],
        "summary": "Get escrow deal by id",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Escrow deal" }
        }
      }
    },
    "/api/escrow/{id}/fund": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Mark escrow as funded (buyer)",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FundEscrowDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Funded" }
        }
      }
    },
    "/api/escrow/{id}/inspection": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Move escrow to inspection",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransitionEscrowDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Inspection started" }
        }
      }
    },
    "/api/escrow/{id}/release": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Release escrow funds to seller",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransitionEscrowDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Released" }
        }
      }
    },
    "/api/escrow/{id}/refund": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Refund escrow to buyer",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransitionEscrowDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Refunded" }
        }
      }
    },
    "/api/escrow/{id}/dispute": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Open an escrow dispute",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DisputeEscrowDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Disputed" }
        }
      }
    },
    "/api/escrow/{id}/cancel": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Cancel an initiated escrow",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransitionEscrowDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Cancelled" }
        }
      }
    }
  }
}